Support running local bundles
This commit adds the ability to configure topaz to run local
policy bundles.

If the policy name passed to `topaz configure` is `"-"` (dash) then
the resource argument is interpreted as a path in the local filesystem.

When starting the authorizer using `topaz start` or `topaz run`, any path
specified in config under `opa.local_bundles.paths` is mounted into the
topaz container using the same path as the host filesystem.