Executable docker image with volume mounts
This changes the topaz Dockerfile to build an image that can be run
as an executable without any command-line arguments.

The entrypoint is set to `/app/topazd run -c` and the default cmd
is `/config/config.yaml`.

When a container is run with no arguments the result is:
`/app/topazd run -c /config/config/yaml`

To override the config file (but not the entry point):
`docker run -v ~/.config/topaz/cfg:/config` topaz:latest /config/conf2.yaml`

To override the entrypoint:
`docker run --entrypoint /app/topaz topaz:latest version`

In addition, the image decalres 4 volume mounts:
* `/config`: location of config file (ro)
* `/certs`: location of certs (ro)
* `/db`: location of database (rw)
* `/decisions`: location of decision logs (rw)