Support custom headers to directory and authorizer
This gives users the ability to configure topaz to include additional
headers (gRPC metadata) in all directory calls.
The remote_directory config value is alread an aserto.Config, which
has a Headers field. This commit simply wires that into the underlying
connection.

It also gives the same ability to the directory and authorizer CLI
commands.