`topaz start` deletes container if it can't be started. (#555)
The call to `ContainerStart` can fail for various reasons including:
* missing or invalid image
* port already in use
* bad volume configuration

All these failures represent a problem at the docker level, not in `topazd`
or its configuration. It's not that the process started and exited with an error.
It never ran at all.

In those cases we should delete the container in order to free up the name. Otherewise
users need to explicitly `docker rm` the container before trying again.